pyproject.toml
[project]
name = "vibe-coding"
version = "0.1.0"
description = ""
authors = [
    { name = "mifrun" }
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "azure-ai-documentintelligence (==1.0.2)",
    "azure-ai-inference (==1.0.0b9)",
    "azure-ai-projects (==2.0.0b2)",
    "azure-identity (==1.25.1)",
    "azure-search-documents (==11.6.0)",
    "email-validator (==2.3.0)",
    "fastapi-cli (==0.0.16)",
    "fastapi-cloud-cli (==0.6.0)",
    "httptools (==0.7.1)",
    "jinja2 (==3.1.6)",
    "mammoth (==1.11.0)",
    "markitdown (==0.1.4)",
    "olefile (==0.47)",
    "openai (==2.11.0)",
    "openpyxl (==3.1.5)",
    "pandas (==2.3.3)",
    "pdfminer-six (==20251107)",
    "playwright (==1.57.0)",
    "pydub (==0.25.1)",
    "python-multipart (==0.0.20)",
    "python-pptx (==1.0.2)",
    "speechrecognition (==3.14.4)",
    "tiktoken (==0.12.0)",
    "uvloop (==0.22.1)",
    "watchfiles (==1.1.1)",
    "xlrd (==2.0.2)",
    "youtube-transcript-api (==1.0.3)",
    "python-dotenv (==1.2.2)",
    "crewai[litellm] (>=1.14.3,<2.0.0)",
]

[tool.poetry]
packages = [{ include = "vibe_coding", from = "src" }]

[project.scripts]
vibe-coding = "vibe_coding.main:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
